[cheri-build] Fix/disable more warnings.

+++ src/heap/mark-compact.cc
@@ -1807,7 +1807,8 @@ class EvacuateVisitorBase : public HeapObjectVisitor {
   bool AbortCompactionForTesting(HeapObject object) {
     if (FLAG_stress_compaction) {
       const uintptr_t mask = static_cast<uintptr_t>(FLAG_random_seed) &
-                             kPageAlignmentMask & ~kObjectAlignmentMask;
+                             kPageAlignmentMask &
+                             ~static_cast<size_t>(kObjectAlignmentMask);
       if ((object.ptr() & kPageAlignmentMask) == mask) {
         Page* page = Page::FromHeapObject(object);
         if (page->IsFlagSet(Page::COMPACTION_WAS_ABORTED_FOR_TESTING)) {
